# Fix #84: Improve sync progress output with batch and daily breakdown