## Summary
This PR improves the sync progress output to provide clearer, more informative progress tracking during conversation syncs.

## Changes
- Added `count_conversations_by_day()` method to IntercomClient to pre-count conversations
- Display daily conversation breakdown before sync starts  
- Show batch/page numbers during sync (e.g., "Batch 4/75")
- Display overall progress percentage across all conversations
- Format ETA in human-readable time (e.g., "5m 30s" instead of "330.0s")
- Update progress line format to be more informative
- Remove confusing per-batch "100%" completion messages

## Example Output

### Before
```
2025-06-28 23:24:51,364 [INFO] Sync progress: 115/115 conversations (100.0%), rate: 5.71/sec, elapsed: 20.1s, ETA: 0.0s
2025-06-28 23:25:13,076 [INFO] Sync progress: 230/230 conversations (100.0%), rate: 5.50/sec, elapsed: 41.8s, ETA: 0.0s
2025-06-28 23:25:36,517 [INFO] Sync progress: 345/345 conversations (100.0%), rate: 5.29/sec, elapsed: 65.3s, ETA: 0.0s
```

### After
```
📊 Analyzing conversations from 2024-12-21 to 2024-12-28...

📅 Conversations to sync:
  • Dec 21: 1,245 conversations
  • Dec 22: 1,189 conversations  
  • Dec 23: 987 conversations
  • Dec 24: 654 conversations
  • Dec 25: 421 conversations
  • Dec 26: 1,456 conversations
  • Dec 27: 1,823 conversations
  • Dec 28: 892 conversations
  📊 Total: 8,667 conversations

⏳ Syncing: 460/8,667 conversations (5.3%) | Batch 4/75 | 5.5 conv/sec | ETA: 24m 52s
⏳ Syncing: 920/8,667 conversations (10.6%) | Batch 8/75 | 5.8 conv/sec | ETA: 22m 18s
```

## Related Issues
Closes #84

## Test Plan
- [x] Unit tests pass
- [x] Code formatting passes (ruff)
- [ ] Integration test with real data shows improved output
- [ ] Performance is not negatively impacted by pre-counting